'writeback' + 'remove' mods as per PR feedback
